Added check to skip duplciate entries based on movie title.

+++ ps-rest-api.php
@@ -1,6 +1,6 @@
 <?php
 /*
-Plugin Name: Horror Movie API Example Plugin v1
+Plugin Name: Horror Movie API Example Plugin
 Plugin URI: https://example.com/
 Description: Example using the HTTP API to parse JSON from a remote horror movie API
 Author: WROX
@@ -63,32 +63,56 @@ function ps_movie_api_results() {
         // Loop through the returned dataset 
         foreach( $data as $movies ) {
 
-            echo '<li>';
-                echo 'Movie imported: ';
-                echo '<a href="https://www.imdb.com/title/' . esc_attr( $movies->imdbId ) . '" target="_blank">';
-                echo esc_html( $movies->title );
-                echo '</a>';
-            echo '</li>';
+            //check if we already have a total stored for today's date
+            echo 'CHECKING MOVIE: ' . esc_html( $movies->title );
 
-            // Set values for our new movies post
-            $new_post = array(
-                'post_title'    => esc_html( $movies->title ),
-                'post_content'  => '<a href="https://www.imdb.com/title/' . esc_attr( $movies->imdbId ) . '">',
-                'post_status'   => 'publish',
-                'post_author'   => 1,
-                'post_type'     => 'movie',
-            );
-            
-            // Insert the post into the database
-            $new_post_id = wp_insert_post( $new_post );
- 
-            // Sideload the image to the media library
-            $image_id = media_sideload_image( esc_url( $movies->posterURL ), $new_post_id, esc_html( $movies->title), 'id' );
+            $existing_movie = get_page_by_title( esc_html( $movies->title ), OBJECT, 'movie' );
 
-            // Set image as featured on the post
-            set_post_thumbnail( $new_post_id, $image_id );
+            if ( !$existing_movie ) {
 
-            //exit();
+                // No entry exists, so let's create one
+
+                echo '<li>';
+                    echo 'Movie imported: ';
+                    echo '<a href="https://www.imdb.com/title/' . esc_attr( $movies->imdbId ) . '" target="_blank">';
+                    echo esc_html( $movies->title );
+                    echo '</a>';
+                echo '</li>';
+
+                // Set values for our new movies post
+                $new_post = array(
+                    'post_title'    => esc_html( $movies->title ),
+                    'post_content'  => '<a href="https://www.imdb.com/title/' . esc_attr( $movies->imdbId ) . '">',
+                    'post_status'   => 'publish',
+                    'post_author'   => 1,
+                    'post_type'     => 'movie',
+                );
+                
+                // Insert the post into the database
+                $new_post_id = wp_insert_post( $new_post );
+     
+                // Sideload the image to the media library
+                $image_id = media_sideload_image( esc_url( $movies->posterURL ), $new_post_id, esc_html( $movies->title), 'id' );
+
+                // Set image as featured on the post
+                set_post_thumbnail( $new_post_id, $image_id );
+
+                //exit();
+
+            }else{
+
+                // Movie entry exists, so we are skipping
+
+                echo '<li>';
+                    echo 'Movie DUPLICATE: ';
+                    echo '<a href="https://www.imdb.com/title/' . esc_attr( $movies->imdbId ) . '" target="_blank">';
+                    echo esc_html( $movies->title );
+                    echo '</a>';
+                echo '</li>';
+
+                //exit();
+
+            }
 
         }
